Rename to spring-rabbit
