+++ instrumentation/spring/spring-rabbit-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/rabbit/MessageHeaderGetter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.spring.amqp;
+package io.opentelemetry.javaagent.instrumentation.spring.rabbit;
 
 import io.opentelemetry.context.propagation.TextMapGetter;
 import org.checkerframework.checker.nullness.qual.Nullable;
