+++ instrumentation/spring/spring-rabbit-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/rabbit/SpringRabbitInstrumentationModule.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.spring.amqp;
+package io.opentelemetry.javaagent.instrumentation.spring.rabbit;
 
 import static java.util.Collections.singletonList;
 
@@ -13,13 +13,13 @@ import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import java.util.List;
 
 @AutoService(InstrumentationModule.class)
-public class SpringAmqpInstrumentationModule extends InstrumentationModule {
-  public SpringAmqpInstrumentationModule() {
-    super("spring-amqp", "spring-amqp-1.0");
+public class SpringRabbitInstrumentationModule extends InstrumentationModule {
+  public SpringRabbitInstrumentationModule() {
+    super("spring-rabbit", "spring-rabbit-1.0");
   }
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new AbstractMessageListenerInstrumentation());
+    return singletonList(new AbstractMessageListenerContainerInstrumentation());
   }
 }
