+++ instrumentation/spring/spring-rabbit-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/spring/rabbit/SpringRabbitSingletons.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.spring.amqp;
+package io.opentelemetry.javaagent.instrumentation.spring.rabbit;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
@@ -11,15 +11,15 @@ import io.opentelemetry.instrumentation.api.instrumenter.SpanNameExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.messaging.MessagingSpanNameExtractor;
 import org.springframework.amqp.core.Message;
 
-public final class SpringAmqpSingletons {
+public final class SpringRabbitSingletons {
 
-  private static final String INSTRUMENTATION_NAME = "io.opentelemetry.javaagent.spring-amqp-1.0";
+  private static final String INSTRUMENTATION_NAME = "io.opentelemetry.javaagent.spring-rabbit-1.0";
 
   private static final Instrumenter<Message, Void> INSTRUMENTER;
 
   static {
-    SpringAmqpMessageAttributesExtractor attributesExtractor =
-        new SpringAmqpMessageAttributesExtractor();
+    SpringRabbitMessageAttributesExtractor attributesExtractor =
+        new SpringRabbitMessageAttributesExtractor();
     SpanNameExtractor<Message> spanNameExtractor =
         MessagingSpanNameExtractor.create(attributesExtractor);
 
@@ -34,5 +34,5 @@ public final class SpringAmqpSingletons {
     return INSTRUMENTER;
   }
 
-  private SpringAmqpSingletons() {}
+  private SpringRabbitSingletons() {}
 }
