+++ settings.gradle
@@ -276,13 +276,13 @@ include ':instrumentation:servlet:servlet-3.0:javaagent'
 include ':instrumentation:servlet:servlet-5.0:library'
 include ':instrumentation:servlet:servlet-5.0:javaagent'
 include ':instrumentation:spark-2.3:javaagent'
-include ':instrumentation:spring:spring-amqp-1.0:javaagent'
 include ':instrumentation:spring:spring-batch-3.0:javaagent'
 include ':instrumentation:spring:spring-core-2.0:javaagent'
 include ':instrumentation:spring:spring-data-1.8:javaagent'
 include ':instrumentation:spring:spring-integration-4.1:javaagent'
 include ':instrumentation:spring:spring-integration-4.1:library'
 include ':instrumentation:spring:spring-integration-4.1:testing'
+include ':instrumentation:spring:spring-rabbit-1.0:javaagent'
 include ':instrumentation:spring:spring-scheduling-3.1:javaagent'
 include ':instrumentation:spring:spring-web-3.1:library'
 include ':instrumentation:spring:spring-webmvc-3.1:javaagent'
